move funcs to helpers
