+++ analysis/process/samples_channel.py
@@ -10,6 +10,7 @@ from pEYES._DataModels.UnparsedEventLabel import UnparsedEventLabelType, Unparse
 from pEYES._DataModels.EventLabelEnum import EventLabelEnum
 
 import analysis.utils as u
+import analysis.process._helpers as h
 
 _CHANNEL_TYPE_STR = "channel_type"
 
@@ -18,7 +19,7 @@ def run_default(
         dataset_name: str,
         pos_labels: Optional[Union[UnparsedEventLabelType, UnparsedEventLabelSequenceType]] = None,
 ):
-    default_output_dir = u.get_default_output_dir(dataset_name)
+    default_output_dir = h.get_default_output_dir(dataset_name)
     try:
         labels = pd.read_pickle(os.path.join(default_output_dir, f"{peyes.constants.LABELS_STR}.pkl"))
     except FileNotFoundError:
@@ -109,7 +110,7 @@ def _calculation_wrapper(
         pos_labels: Optional[Union[UnparsedEventLabelType, UnparsedEventLabelSequenceType]] = None,
         iteration_desc: str = "Channel Metrics",
 ) -> dict:
-    pred_labelers = u.check_labelers(labels, pred_labelers)
+    pred_labelers = h.check_labelers(labels, pred_labelers)
     if pos_labels is None:
         pos_labels = [l for l in EventLabelEnum]
     elif isinstance(pos_labels, UnparsedEventLabelType):
